Move new app pytest fixtures to conftest from test_asmts_workflow.py
